Refactor Deepchecks components and update dependencies structure
Split and reorganize Deepchecks components into separate modules (`dc_data` and `dc_model`) for improved structure and clarity. Updated import paths across the codebase accordingly. Adjusted dependencies in `uv.lock` and Makefile to reflect the new changes and added `black` for consistent code formatting.
