+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        python-version: [ '3.9', '3.10', '3.11', '3.12' ]
+        python-version: [ '3.9', '3.10',  ]
     steps:
       - uses: actions/checkout@v4
       - name: Install OS Dependencies
@@ -23,8 +23,7 @@ jobs:
       - uses: astral-sh/setup-uv@v4
         with:
           enable-cache: true
-      - run: uv python install ${{ matrix.python-version }}
-      - run: uv sync --all-extras --frozen
+      - run: make
       - run: make test
     continue-on-error: true
   tests-imports:
@@ -40,10 +39,10 @@ jobs:
       - uses: astral-sh/setup-uv@v4
         with:
           enable-cache: true
-      - run: uv python install 3.11
+      - run: uv python install 3.10
       - run: uv sync --extra ${{ matrix.extra }} --frozen
       - run: uv run pytest tests/imports_checks/test_${{ matrix.extra }}_imports.py
-      - name: test-mljar-imports
-        if: matrix.extra == 'mljar'
-        run: uv run pytest tests/comp/mljar
+#      - name: test-mljar
+#        if: matrix.extra == 'mljar'
+#        run: uv run pytest tests/comp/mljar
     continue-on-error: true
