+++ .github/workflows/publish.yml
@@ -3,9 +3,9 @@ name: Publish Package
 on:
   release:
     types: [ published ]
-  pull_request:
-    branches:
-      - main
+#  pull_request:
+#    branches:
+#      - main
 
 
 permissions:
