+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: help
 
 install:
-	uv sync --all-extras --frozen
+	uv sync --all-extras --all-groups --frozen
 	uv tool install pre-commit --with pre-commit-uv --force-reinstall
 
 .PHONY: default
