+++ pyproject.toml
@@ -30,9 +30,9 @@ editor = [
 ]
 
 
-#evalml = [
-#    "evalml>=0.83.0","ml-orchestrator", "pandas-pyarrow"
-#]
+evalml = [
+    "evalml>=0.83.0","ml-orchestrator", "pandas-pyarrow"
+]
 
 spliter = [
     "scikit-learn>=1.3.0", "ml-orchestrator", "pandas-pyarrow"
