+++ tabular_kfp/parse_components.py
@@ -1,7 +1,7 @@
 from ml_orchestrator import ComponentParser
 
 from tabular_orchestrated.components import DataSplitter
-from tabular_orchestrated.deepchecks import DCDataComp, DCTrainTestComp
+from tabular_orchestrated.dc.dc_data import DCDataComp, DCTrainTestComp
 from tabular_orchestrated.mljar.mljar import EvaluateMLJAR, MLJARTraining
 from tabular_orchestrated.mljar.mljar_deepchecks import MljarDCFullComp, MljarDCModelComp
