+++ tabular_orchestrated/dc/dc_data.py
@@ -1,0 +1,33 @@
+import dataclasses
+from typing import Any
+
+from deepchecks.tabular.suites import data_integrity, train_test_validation
+from ml_orchestrator import artifacts
+from ml_orchestrator.artifacts import Input
+
+from tabular_orchestrated.dc.dc_main import DCMetaComp
+
+
+@dataclasses.dataclass
+class DCDataComp(DCMetaComp):
+    dataset: Input[artifacts.Dataset] = None
+
+    def prepare_suite(self) -> Any:
+        data = self.load_df(self.dataset)
+        suite = data_integrity()
+        dc_data = self.transform_dataframe(data)
+        return suite.run(dc_data)
+
+
+@dataclasses.dataclass
+class DCTrainTestComp(DCMetaComp):
+    train_dataset: Input[artifacts.Dataset] = None
+    test_dataset: Input[artifacts.Dataset] = None
+
+    def prepare_suite(self) -> Any:
+        train_data = self.load_df(self.train_dataset)
+        test_data = self.load_df(self.test_dataset)
+        suite = train_test_validation()
+        dc_train_dataset = self.transform_dataframe(train_data)
+        dc_test_dataset = self.transform_dataframe(test_data)
+        return suite.run(dc_train_dataset, dc_test_dataset)
