+++ tests/comp/dc/test_comp_run.py
@@ -3,22 +3,8 @@ from pathlib import Path
 from ml_orchestrator import artifacts
 
 from tabular_orchestrated.components import DataSplitter
-from tabular_orchestrated.deepchecks import DCDataComp, DCTrainTestComp
-
-#
-# def test_mljar(get_df_example: artifacts.Dataset, split_op: DataSplitter,
-#                deepchecks_op: DCFullComp, deepchecks_data_op: DCDataComp,
-#                deepchecks_train_test_op: DCTrainTestComp, ) -> None:
-#     tmp_files_folder = Path(get_df_example.uri).parent
-#
-#     def func(x):
-#         return (tmp_files_folder / x).as_posix()
-#
-#     assert Path(func("train.parquet")).exists()
-#     assert Path(func("test.parquet")).exists()
-#     assert Path(func("deepchecks.html")).exists()
-#     assert Path(func("deepchecks_data.html")).exists()
-#     assert Path(func("deepchecks_train_test.html")).exists()
+from tabular_orchestrated.dc.dc_data import DCDataComp
+from tabular_orchestrated.dc.dc_model import DCTrainTestComp
 
 
 def test_split(get_df_example: artifacts.Dataset, split_op: DataSplitter) -> None:
