+++ tests/imports_checks/test_deepchecks_imports.py
@@ -1,4 +1,4 @@
 def test_dc_imports():
-    from tabular_orchestrated.deepchecks import DCTrainTestComp, DCDataComp  # noqa
+    from tabular_orchestrated.dc import DCTrainTestComp, DCDataComp  # noqa
 
     assert True
